requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-mongodb-engine==0.6.0
djangorestframework==3.13.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
filelock==3.6.0
flake8==4.0.1
jedi==0.18.1
mccabe==0.6.1
mongoengine==0.24.1
numpy==1.22.3
pandas==1.4.1
parso==0.8.3
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
termcolor==1.1.0
tqdm==4.63.1
virtualenv==20.14.0
